Task 2.8: Implemented orderPageController Implemented OrderService Added new methods to phoneDao Added new methods to cartService Added some new tests Code refactoring: moved some directories / packages

+++ core/src/main/java/com/es/core/model/phone/Stock.java
@@ -2,8 +2,8 @@ package com.es.core.model.phone;
 
 public class Stock {
     private Phone phone;
-    private Integer stock;
-    private Integer reserved;
+    private Long stock;
+    private Long reserved;
 
     public Phone getPhone() {
         return phone;
@@ -13,19 +13,19 @@ public class Stock {
         this.phone = phone;
     }
 
-    public Integer getStock() {
+    public Long getStock() {
         return stock;
     }
 
-    public void setStock(Integer stock) {
+    public void setStock(Long stock) {
         this.stock = stock;
     }
 
-    public Integer getReserved() {
+    public Long getReserved() {
         return reserved;
     }
 
-    public void setReserved(Integer reserved) {
+    public void setReserved(Long reserved) {
         this.reserved = reserved;
     }
 }
